backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6

# HTTP and async
httpx==0.25.1
aiofiles==23.2.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.2
alembic==1.12.1

# Geo libraries
geopy==2.4.0
shapely==2.0.2
pyproj==3.6.1
geopandas==0.14.1

# Data processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Caching
redis==5.0.1
hiredis==2.2.3

# Logging
python-json-logger==2.0.7

# Resilience patterns
tenacity==8.2.3
py-circuit-breaker==0.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.25.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# Monitoring (optional)
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0